agents/architect/requirements.txt
fastapi
uvicorn
pika
redis
google-generativeai
